add send mail util and started repeat send controller
