+++ controller/users.js
@@ -2,6 +2,7 @@ const { User } = require("../service/schemas/user.js");
 const service = require("../service/users.js");
 const { avatarDir } = require("../middlewares/upload.js");
 const { editAvatar } = require("../utils/editAvatar.js");
+const { sendMail } = require("../utils/sendVerifyMail.js");
 const sgMail = require("@sendgrid/mail");
 const { v4: uuidv4 } = require("uuid");
 const jwt = require("jsonwebtoken");
@@ -24,23 +25,12 @@ const getAll = async (req, res, next) => {
 
 const register = async (req, res, next) => {
   const { email, password } = req.body;
-  sgMail.setApiKey(sgToken);
   const user = await service.getUser({ email });
 
   if (user) return res.status(409).json({ message: "Email in use" });
-
   const vfToken = uuidv4();
-  const vfLink = `http://localhost:3000/api/users/verify/${vfToken}`;
-  const msg = {
-    to: email,
-    from: "mlody13992@gmail.com",
-    subject: "Verification token",
-    text: `Your verification token: ${vfLink}`,
-    html: `<b>Your verification token: <a href="${vfLink}">${vfLink}</a></b>`,
-  };
-
   try {
-    await sgMail.send(msg);
+    await sendMail(email, vfToken);
     const avatarURL = gravatar.url(email, { s: "250", d: "mp" });
     const newUser = new User({ email, avatarURL, verificationToken: vfToken });
     newUser.setPassword(password);
@@ -63,6 +53,10 @@ const login = async (req, res, next) => {
 
     if (!user || !user.validPassword(password))
       return res.status(401).json({ message: "Email or password is wrong" });
+    if (!user.verify)
+      return res.status(401).json({
+        message: "User is not verified, please click verification link",
+      });
 
     const payload = {
       id: user.id,
@@ -143,6 +137,21 @@ const verificationLink = async (req, res, next) => {
       { new: true }
     );
     res.status(200).json({ message: "Verification successful" });
+  } catch (err) {
+    next(err);
+  }
+};
+
+const repeatVerification = async (req, res, next) => {
+  try {
+    const { email } = req.body;
+    if (!email)
+      return res.status(400).json({ message: "missing required field email" });
+    const user = await service.getUser({ email });
+    if (user.verify)
+      return res
+        .status(400)
+        .json({ message: "Verification has already been passed" });
   } catch (err) {
     next(err);
   }
